Removed streambuf silliness in stream_test. Now stream test is built differently and only takes strings. If client wants to use a file, they can use ChunkReader. So then the plan is to make an 'factory' of some sorts that returns the either a) or b).
This in turn makes AReader not a pain-in-the-rear to init.
Also yet to change json_adt/main.cpp to reflect these changes.

+++ chunkreader/ireader.h
@@ -1,6 +1,5 @@
 #ifndef IREADER
 #define IREADER
-
 #include <string>
 
 using std::string;
@@ -9,7 +8,7 @@ class IReader {
 	protected:
 		bool good;
 		const size_t M;
-		explicit IReader(const int chars): good(true), M(chars){}
+		explicit IReader(const size_t chars): good(true), M(chars){}
 	public:
 		virtual ~IReader(){}
 		virtual IReader& operator++() = 0;
@@ -20,5 +19,4 @@ class IReader {
 		virtual char until(char c,string* str_ptr=nullptr) = 0;
 		virtual bool empty() = 0;
 };
-
 #endif
